cleanup: move a number of functions from qthelper.cpp to string-format.cpp
This is a long-term project in detangling qthelper.cpp, which was started
many moons ago.

The formatting functions are a royal mess, but collecting them in one
place seems like a start.
