+++ desktop-widgets/tab-widgets/TabDiveNotes.cpp
@@ -5,6 +5,7 @@
 #include "core/qthelper.h"
 #include "core/selection.h"
 #include "core/subsurface-string.h"
+#include "core/string-format.h"
 #include "core/tag.h"
 #include "core/trip.h"
 #include "desktop-widgets/mainwindow.h"
